fix center calculation when clicking on a QuickItem
The previous version is correct only if the widget orientation is not
modified. A better solution is then to pass the center of the widget
as mapToScene() argument (this argument being in widget's coordinate
system) then let this method mapping the point to scene coordinate
(which is done by taking into account transformations applied to the
item).