add drag'n'drop support for QuickView UI
A new QuickDragNDropResponse class is created, based on
DragNDropResponse, to support drag'n'drop on QuickView based UI. This
is done this way because the existing class is only compatible with
QWidget (which is not inherited by QQuickItem).

Also, intermediate positions injection during mouse press are spaced by
a few milliseconds instead of posting all of them at once in the
QApplication. This change gives better results with some elements (like
SwipeView) by producing a more natural movement. For this to be
implemented, m_nbCall member variable of DelayedResponse has been made
visible by subclasses.