Server: Pass build type to "make" instead of "qmake"
With a recent Qt version, building funq-server failed on Windows with
following error:

    No rule to make target 'release'

By passing the build type (debug/release) only to the "make" command,
but not to the "qmake" command fixes the error. Builds on Linux and/or
with older Qt versions are not affected by this change, according CI
they still work as expected.